+++ src/Exceptions/ServiceException.php
@@ -3,11 +3,12 @@
 namespace Honeybadger\Exceptions;
 
 use Exception;
+use Throwable;
 
 class ServiceException extends Exception
 {
     /**
-     * @return \Honeybadger\Exceptions\ServiceException
+     * @return ServiceException
      */
     public static function invalidApiKey(): self
     {
@@ -15,7 +16,7 @@ class ServiceException extends Exception
     }
 
     /**
-     * @return \Honeybadger\Exceptions\ServiceException
+     * @return ServiceException
      */
     public static function invalidPayload(): self
     {
@@ -23,7 +24,7 @@ class ServiceException extends Exception
     }
 
     /**
-     * @return \Honeybadger\Exceptions\ServiceException
+     * @return ServiceException
      */
     public static function rateLimit(): self
     {
@@ -31,24 +32,49 @@ class ServiceException extends Exception
     }
 
     /**
-     * @return \Honeybadger\Exceptions\ServiceException
+     * @return ServiceException
      */
     public static function serverError(): self
     {
         return new static('There was an error on our end.');
     }
 
+    /**
+     * @param int $code
+     * @return ServiceException
+     */
     public static function unexpectedResponseCode(int $code): self
     {
         return new static("Unexpected HTTP response code: $code");
     }
 
-    public static function generic(\Throwable $e = null): self
+    /**
+     * @param Throwable|null $e
+     * @return self
+     */
+    public static function generic(Throwable $e = null): self
     {
         $message = $e
             ? 'There was an error sending the payload to Honeybadger: '.$e->getMessage()
             : 'There was an error sending the payload to Honeybadger.';
 
         return new static($message, 0, $e);
+    }
+
+    /**
+     * @param string $message
+     * @return self
+     */
+    public static function invalidConfig(string $message): self
+    {
+        return new static("The configuration is invalid: $message");
+    }
+
+    /**
+     * @return self
+     */
+    public static function missingPersonalAuthToken(): self
+    {
+        return new static("Missing personal auth token. This token is required to use Honeybadger's Data APIs.");
     }
 }